requirements.txt
tensorflow-gpu >= 1.8
numpy >= 1.11.1
librosa == 0.5.1
joblib == 0.11.0
tensorpack >= 0.8.6
pyyaml
soundfile
pydub
tqdm
git+https://github.com/wookayin/tensorflow-plot.git@master